// src/modules/reports/dto/report-interfaces.ts
export interface ITabularField {
  id: string;
  draggedId: string;
  columnName?: string;
  columnDisplayName: string;
  type: string;
  operation?: string;
  node?: string;
  isExplodedBy?: boolean;
}

export interface IReportField extends ITabularField {
  footerAggregation: Array<string>;
  decimalNumbers?: number;
  thresholdUpperValue?: number;
  thresholdLowerValue?: number;
  dateFormat?: string;
  hidden: boolean;
  pinned: boolean;
  trunc?: boolean;
  round?: boolean;
  index: number;
  trValue?: number;
  selected?: boolean;
  isMetric?: boolean;
}

export interface IMinimalTabularTable {
  id: string;
  displayName: string;
  role?: string;
  fields: Array<IReportField>;
}

export interface ITabularTable extends IMinimalTabularTable {
  tableName: string;
  startTime: string;
  statInterval: number;
  tableHourName: string;
  tableDayName: string;
  paramsTable: string;
  paramsNodeName: string;
  nodeNameColumn: string;
  statDateNameColumn: string;
  gracePeriod: number;
}

// --- Order By ---

export interface ITabularOrderBy {
  draggedId: string;
  columnDisplayName: string;
  orderBy: string;
}
// ...
export interface IOperationColumn {
  isOperation: boolean;
  isCustomColumn: boolean;
  draggedId: string;
}

export interface IOperationOperatorColumn {
  isOperation: boolean;
  operator: string;
}

export type OperationColumnType = IOperationOperatorColumn | IOperationColumn;
// ...
export interface ICustomOperationColumn {
  draggedId: string;
  columnDisplayName: string;
  footerAggregation: Array<string>;
  type: string;
  operation: string;
  savedTokens: Array<OperationColumnType>;
  customColumn: boolean;
  customColumnType: string;
  hidden: boolean;
  pinned: boolean;
  index: number;
  trunc: boolean;
  round: boolean;
  trValue: number;
  selected?: boolean;
  isMetric?: boolean;
  isExplodedBy?: boolean;
}
// ...
export interface IChartData {
  id: string;
  name: string;
  type: string;
  orderIndex: number;
  [key: string]: unknown;
}

// --- Tabular Header ---

export interface ITabularHeader {
  text: string;
  datafield: string;
  columnName?: string;
  aggregates: string[];
  pinned: boolean;
  hidden: boolean;
  draggedId?: string;
  editable?: boolean;
  columntype?: string;
  index?: number;
  headerColumnType?: string;
}